+++ src/ui/styled/globals/data/index.ts
@@ -1,0 +1,8 @@
+import type { GlobalStyleObject } from '@pandacss/types';
+import { fontSizeDataStyles } from './fontSizes';
+import { spacingDataStyles } from './spacing';
+
+export const dataCss: GlobalStyleObject = {
+  ...spacingDataStyles,
+  ...fontSizeDataStyles,
+};
